reduce gpu memory usage option : discrete chain

+++ src/platforms/cpu/CpuPseudoContinuous.cpp
@@ -199,7 +199,7 @@ void CpuPseudoContinuous::compute_statistics(
                 // if it is not leaf node
                 else if (n_segment_from == 1 && deps.size() > 0) 
                 {
-                    // if it is superposition
+                    // if it is superposed
                     if (key[0] == '[')
                     {
                         for(int i=0; i<M; i++)
@@ -333,7 +333,7 @@ void CpuPseudoContinuous::compute_statistics(
 
         // calculate segment concentrations
         #pragma omp parallel for
-        for(int b=0; b<unique_phi.size();b++)
+        for(size_t b=0; b<unique_phi.size();b++)
         {
             auto block = unique_phi.begin();
             advance(block, b);
@@ -572,7 +572,7 @@ std::vector<double> CpuPseudoContinuous::compute_stress()
 
         // compute stress for each unique block
         #pragma omp parallel for
-        for(int b=0; b<unique_phi.size();b++)
+        for(size_t b=0; b<unique_phi.size();b++)
         {
             auto block = unique_phi.begin();
             advance(block, b);
